Fix intellij-buf build w/ new buf versions
Instead of installing the latest license-header binary, we should pin it
to the same version as the version of Buf we're using. Upgrade buf to
the latest v1.22.0 release, and relax the restriction in the build that
we need an Azul toolchain - any Java 11 toolchain should work until we
upgrade to Java 17.