Reduce auto-save impact for lint/breaking checks (#136)
Currently, any editor with a .proto file open will trigger reformatting
(even if the .proto file itself isn't changed). This appears to be
related to changes in editor focus when background jobs are run.

Update the plugin to keep track of a modification count based on the
protobuf files in the project in addition to the Buf CLI configuration
and plugin configuration. Only run 'buf breaking' and 'buf lint' if the
count changes (since changes to non-proto/buf files won't impact the
results).

This should greatly reduce the times we run background jobs and avoid
unnecessary reformatting of non-proto files during editor sessions.

Fixes #133.