Preserve newlines with buf format (#181)
The previous implementation of formatting would take each chunk of
stdout (up to 8192 chars), trim any trailing newlines, and then join the
chunks together with a newline at the end. This causes short inputs to
be missing a trailing newline at EOF and also in some cases caused
newlines to be inserted in the middle of chunks of a protobuf file
causing syntax errors.

Update the plugin to stop trimming the end of stdout chunks and join
without newlines, preserving output exactly as-is from buf format.
Update the tests to add a new larger file and fix the expected result of
the first test to add the trailing newline at EOF. Now both inputs match
the output of 'buf format' exactly.

Fixes #180.