Fix logging of expected exit codes (#216)
Both the lint and breaking change command are expected to return exit
code 100 when issues are found. Update the logging to omit logging these
cases so we only log warnings if a real failure occurs trying to run the
Buf CLI.

Additionally, add a comment regarding file-level lint warnings (which
are currently unsupported).