Fix linting in IntelliJ plugin
Starting in the 2025.1 release, long deprecated APIs relied on by the
analysis pass are no longer functioning. Migrate off AnnotationImpl and
AnnotationHolder to use HighlightInfo.

As part of this change, we're going to bump the minor version of the
plugin and target 2025.1 as the min supported version. Also upgrade to
the latest Gradle and IntelliJ gradle plugin.